move TextArea to common
