+++ src/__test__/snapShotComponents.test.tsx
@@ -3,9 +3,8 @@ import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import renderer from 'react-test-renderer';
 
-import { Button, Input, Form, Heading, Message, Topic } from '../components';
+import { Button, Input, Form, Heading, Message, Topic, TextArea } from '../components';
 
-import TextArea from '../components/TextArea';
 import InfoPanel from '../components/MainPageComponents/InfoPanel';
 import MainMenuItem from '../components/MainPageComponents/MainMenuItem';
