+++ src/components/common/TextArea/index.ts
@@ -1,0 +1,1 @@
+export * from './TextArea';
