+++ src/components/common/index.ts
@@ -1,10 +1,11 @@
 // такой индекс-файл нужен для того чтобы можно было импортировать
 // компонент просто указав путь до папки components
 
-export * from './Modal';
 export * from './Button';
 export * from './ErrorBoundary';
-export * from './Input';
 export * from './Form';
 export * from './Heading';
+export * from './Input';
+export * from './Modal';
 export * from './Scroll';
+export * from './TextArea';
